backend/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
pandas==1.5.1
python-dotenv==0.21.0
tweepy==4.12.1
